Conflicts:
	Unity Project/Assets/Scenes/KC Scene.unity

Former-commit-id: 4b7ca12ad78f775e4952312517cb66ae87247e5a

+++ Unity Project/Assets/Scripts/Common/Probability/ProbabilityLevItem.cs
@@ -3,5 +3,5 @@ using System.Collections.Generic;
 
 public interface ProbabilityLevItem : ProbabilityItem
 {
-    int Level { get; }
+    ushort Level { get; }
 }
